+++ airflow/contrib/example_dags/example_azure_container_instances_operator.py
@@ -22,7 +22,9 @@ This is an example dag for using the AzureContainerInstancesOperator.
 from datetime import datetime, timedelta
 
 from airflow import DAG
-from airflow.contrib.operators.azure_container_instances_operator import AzureContainerInstancesOperator
+from airflow.providers.microsoft.azure.operators.azure_container_instances import (
+    AzureContainerInstancesOperator,
+)
 
 default_args = {
     'owner': 'airflow',
